+++ packages/ts-utils/src/string-sorting.ts
@@ -1,29 +1,27 @@
 /**
- * Various functional utilities for sorting things.
+ * Various functional utilities for sorting things with strings.
  */
 
-
-type Comparer<T> = (l: T, r: T) => number
-
-const stringCompare: Comparer<string> = (l, r): number =>
-    l === r ? 0 : (l > r ? 1 : -1)
-
-const stringyCompare = <T>(keyFunc: (t: T) => string): Comparer<T> =>
-    (l: T, r: T) => stringCompare(keyFunc(l), keyFunc(r))
-
-export const sortByStringKey = <T>(ts: T[], keyFunc: (t: T) => string) =>
-    [...ts].sort(stringyCompare(keyFunc))
+import { sorterWith } from "./sorting.js"
+import { localeStringComparer, mappedComparer, regularStringComparer } from "./comparer.js"
 
 
+/**
+ * @deprecated Use a {@link Sorter} built with {@link sorterWith}, instead.
+ * Build the required {@link Comparer} using {@link }
+ */
+export const sortByStringKey = <T>(ts: T[], stringFunction: (t: T) => string) =>   // TODO  rename -> sortByStringFunction
+    sorterWith(mappedComparer(stringFunction, regularStringComparer))(ts)
+
+
+/**
+ * @deprecated Use `Sorter<string>` instead.
+ */
 export type StringSorter = (strings: string[]) => string[]
 
-const sortedStringsWith = (strMap: (str: string) => string): StringSorter =>
-    (strings) => {
-        strings.sort((l, r) => strMap(l).localeCompare(strMap(r)))
-        return strings
-    }
+export const sortedStrings = sorterWith(localeStringComparer)
 
-export const sortedStrings = sortedStringsWith((str) => str)
-
-export const sortedStringsByUppercase = sortedStringsWith((str) => str.toUpperCase())
+export const sortedStringsByUppercase = sorterWith<string>(
+    mappedComparer((str) => str.toUpperCase(), localeStringComparer)
+)