+++ packages/ts-utils/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Implement a small library for lexicographically comparisons (of strings, mainly): `Comparer` type, `{regular|locale}StringComparer` constants, and `mappedComparer`, `lexiComparer`, and `mappedLexiComparer` functions.
 * Expose a `Sorter` type, and `sorterWith` function for sorting.
+* Deprecate `StringSorter` type, `sortByStringKey` function, and replace its uses.
 
 
 ## 0.9.0